changed int types to size_t

+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: gguedes <gguedes@student.42.rio>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/14 12:12:52 by gguedes           #+#    #+#             */
-/*   Updated: 2022/05/14 12:30:56 by gguedes          ###   ########.fr       */
+/*   Updated: 2022/05/14 12:40:07 by gguedes          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *s, int c, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
